+++ grako/grammars.py
@@ -452,10 +452,10 @@ class PositiveClosure(Closure):
 
 
 class Join(Decorator):
-    def __init__(self, ast=None, keep_sep=False, **kwargs):
+    def __init__(self, ast=None, keepsep=False, **kwargs):
         super(Join, self).__init__(ast.exp)
         self.sep = ast.sep
-        self.keep_sep = keep_sep
+        self.keepsep = keepsep
 
     def parse(self, ctx):
         def sep():
@@ -464,10 +464,10 @@ class Join(Decorator):
         def exp():
             return self.exp.parse(ctx)
 
-        return self._closure(ctx, exp, sep, self.keep_sep)
+        return self._closure(ctx, exp, sep, self.keepsep)
 
-    def _closure(self, ctx, exp, sep, keep_sep):
-        return ctx._closure(exp, sep=sep, keep_sep=keep_sep)
+    def _closure(self, ctx, exp, sep, keepsep):
+        return ctx._closure(exp, sep=sep, keepsep=keepsep)
 
     def _to_str(self, lean=False):
         ssep = self.sep._to_str(lean=lean)
@@ -479,8 +479,8 @@ class Join(Decorator):
 
 
 class PositiveJoin(Join):
-    def _closure(self, ctx, exp, sep, keep_sep):
-        return ctx._positive_closure(exp, sep=sep, keep_sep=keep_sep)
+    def _closure(self, ctx, exp, sep, keepsep):
+        return ctx._positive_closure(exp, sep=sep, keepsep=keepsep)
 
     def _to_str(self, lean=False):
         return super(PositiveJoin, self)._to_str(lean=lean) + '+'
@@ -488,12 +488,12 @@ class PositiveJoin(Join):
 
 class KeeperJoin(Join):
     def __init__(self, ast=None, **kwargs):
-        super(KeeperJoin, self).__init__(ast=ast, keep_sep=True, **kwargs)
+        super(KeeperJoin, self).__init__(ast=ast, keepsep=True, **kwargs)
 
 
 class KeeperPositiveJoin(PositiveJoin):
     def __init__(self, ast=None, **kwargs):
-        super(KeeperPositiveJoin, self).__init__(ast=ast, keep_sep=True, **kwargs)
+        super(KeeperPositiveJoin, self).__init__(ast=ast, keepsep=True, **kwargs)
 
 
 class EmptyClosure(Model):